requirements.txt
anthropic>=0.40.0
httpx>=0.26.0
resend>=2.0.0
rich>=13.0
python-dotenv>=1.0.0
pyyaml>=6.0
fastapi>=0.115
uvicorn>=0.30
python-multipart>=0.0.12
jinja2>=3.1
aiosqlite>=0.20.0
